blargh

+++ CemeteryFunTimes/src/cemeteryfuntimes/Resources/Shared/Collision.java
@@ -2,10 +2,9 @@ package cemeteryfuntimes.Resources.Shared;
 
 // @author David Kozloff & Tyler Law
 
-
 public class Collision implements Globals {
     
-    public static int checkWallCollision(float xPos, int xRad, float yPos, int yRad) {
+    public static int checkWallCollision(float xPos, float xRad, float yPos, float yRad) {
         if (xPos + xRad > GAMEWIDTH) {
             return RIGHTWALL;
         }
@@ -20,5 +19,4 @@ public class Collision implements Globals {
         }
         return 0;
     }
-    
 }
